feat: updating to docusaurus v3
