+++ package.json
@@ -24,24 +24,24 @@
   "devDependencies": {
     "@cloudflare/kv-asset-handler": "^0.2.0",
     "@cloudflare/workers-types": "^3.14.1",
-    "@docusaurus/core": "^2.3.1",
-    "@docusaurus/module-type-aliases": "^2.3.1",
-    "@docusaurus/plugin-client-redirects": "^2.3.1",
-    "@docusaurus/plugin-content-docs": "^2.3.1",
-    "@docusaurus/plugin-debug": "^2.3.1",
-    "@docusaurus/plugin-google-analytics": "^2.3.1",
-    "@docusaurus/plugin-google-gtag": "^2.3.1",
-    "@docusaurus/theme-classic": "^2.3.1",
-    "@docusaurus/theme-common": "^2.3.1",
-    "@docusaurus/theme-search-algolia": "^2.3.1",
-    "@docusaurus/types": "^2.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.55.0",
-    "eslint-config-prettier": "^8.6.0",
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@docusaurus/plugin-client-redirects": "^3.0.0",
+    "@docusaurus/plugin-content-docs": "^3.0.0",
+    "@docusaurus/plugin-debug": "^3.0.0",
+    "@docusaurus/plugin-google-gtag": "^3.0.0",
+    "@docusaurus/plugin-google-tag-manager": "^3.0.0",
+    "@docusaurus/theme-classic": "^3.0.0",
+    "@docusaurus/theme-common": "^3.0.0",
+    "@docusaurus/theme-search-algolia": "^3.0.0",
+    "@docusaurus/types": "^3.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.61.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^4.2.1",
-    "prettier": "^2.8.4",
-    "typescript": "^4.5.2",
-    "unist-util-visit": "^2.0.3",
+    "eslint-plugin-prettier": "^5.0.0-alpha.2",
+    "prettier": "^3.0.0",
+    "typescript": "^5.1.6",
+    "unist-util-visit": "^5.0.0",
     "wrangler": "^2.0.23"
   }
 }
