+++ server/tsconfig.json
@@ -2,6 +2,7 @@
 // It is only used for IDE checking and auto-completion
 {
   "compilerOptions": {
+    "composite": true,
     "allowJs": true,
     "strictNullChecks": true,
     "noImplicitAny": false,
@@ -18,7 +19,6 @@
       "node",
       "@cloudflare/workers-types",
     ],
-    "noEmit": true,
   },
   "include": [
     "./**/*"
