+++ sidebars.ts
@@ -1,7 +1,6 @@
-// @ts-check
+import type {SidebarsConfig} from '@docusaurus/plugin-content-docs';
 
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-const sidebars = {
+const sidebars: SidebarsConfig = {
   docs: [
     {
       type: "category",
@@ -193,4 +192,4 @@ const sidebars = {
   ],
 };
 
-module.exports = sidebars;
+export default sidebars;
